requirements.txt
gradio>=4.0.0
torch>=2.0.0
transformers>=4.35.0
anthropic>=0.18.0
google-genai>=0.2.0
openai>=1.0.0
biopython>=1.81
plotly>=5.18.0
numpy>=1.24.0
sentencepiece>=0.1.99
protobuf>=3.20.0
huggingface-hub>=0.19.0
git+https://github.com/McClain-Thiel/plasmid-kit.git
